♻️ Refactor: enum 공통 관리 수정 (#29)
